split out views into app detail and add more to base views, update urls
