+++ intake/views/base_views.py
@@ -1,3 +1,5 @@
+from django.contrib.auth.mixins import PermissionRequiredMixin
+from intake import permissions
 import intake.services.counties as CountiesService
 
 
@@ -10,3 +12,7 @@ class GlobalTemplateContextMixin:
             all_county_names='counties throughout California',
             organizations=orgs
         )
+
+
+class ViewAppDetailsMixin(PermissionRequiredMixin):
+    permission_required = permissions.CAN_SEE_APP_DETAILS.app_code
